app/requirements.txt
Flask
Redis
python-dotenv
Pandas